+++ src/datajudge/constraints/miscs.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import List, Optional, Set, Tuple
 
 import sqlalchemy as sa
@@ -8,8 +9,8 @@ from .base import Constraint, OptionalSelections, TestResult, format_sample
 
 
 class PrimaryKeyDefinition(Constraint):
-    def __init__(self, ref, primary_keys: List[str]):
-        super().__init__(ref, ref_value=set(primary_keys))
+    def __init__(self, ref, primary_keys: List[str], name: str = None):
+        super().__init__(ref, ref_value=set(primary_keys), name=name)
 
     def retrieve(
         self, engine: sa.engine.Engine, ref: DataReference
@@ -53,6 +54,8 @@ class Uniqueness(Constraint):
         ref: DataReference,
         max_duplicate_fraction: float = 0,
         max_absolute_n_duplicates: int = 0,
+        infer_pk_columns: bool = False,
+        name: str = None,
     ):
         if max_duplicate_fraction != 0 and max_absolute_n_duplicates != 0:
             raise ValueError(
@@ -66,9 +69,24 @@ class Uniqueness(Constraint):
             ref_value = ("absolute", max_absolute_n_duplicates)
         else:
             ref_value = ("relative", 0)
-        super().__init__(ref, ref_value=ref_value)
+
+        self.infer_pk_columns = infer_pk_columns
+        super().__init__(ref, ref_value=ref_value, name=name)
 
     def test(self, engine: sa.engine.Engine) -> TestResult:
+
+        # only check for primary keys when actually defined
+        # otherwise default back to searching the whole table
+        if self.infer_pk_columns and (
+            pk_columns := db_access.get_primary_keys(engine, self.ref)[0]
+        ):
+            self.ref.columns = pk_columns
+            if not pk_columns:  # there were no primary keys found
+                warnings.warn(
+                    f"""No primary keys found in {self.ref.get_string()}.
+                    Uniqueness will be tested for all columns."""
+                )
+
         unique_count, unique_selections = db_access.get_unique_count(engine, self.ref)
         row_count, row_selections = db_access.get_row_count(engine, self.ref)
         self.factual_selections = row_selections
@@ -98,10 +116,10 @@ class Uniqueness(Constraint):
 
 
 class NullAbsence(Constraint):
-    def __init__(self, ref: DataReference):
+    def __init__(self, ref: DataReference, name: str = None):
         # This is arguably hacky. Passing this pointless string ensures that
         # None-checks fail.
-        super().__init__(ref, ref_value="NoNull")
+        super().__init__(ref, ref_value="NoNull", name=name)
 
     def test(self, engine: sa.engine) -> TestResult:
         assertion_message = f"{self.ref.get_string()} contains NULLS."